requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
Pillow>=10.2.0
openai>=1.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
